feat: implement PasswordResetForm component with validation and styling
