+++ src/presentation/widgets/auth/PasswordResetForm/PasswordResetForm.module.css
@@ -1,20 +1,20 @@
-.form-flex {
+.formFlex {
   display: flex;
   flex-direction: column;
   flex: 1 1 auto;
 }
 
-.form-fields-container {
+.formFieldsContainer {
   display: flex;
   flex-direction: column;
   gap: 16px;
 }
 
-.change-password-btn {
+.changePasswordBtn {
   width: 100%;
 }
 
-.btn-container {
+.btnContainer {
   margin-top: auto;
   padding-bottom: calc(var(--bottom-nav-height) + 20px);
 }
