+++ src/presentation/widgets/auth/PasswordResetForm/index.ts
@@ -1,0 +1,1 @@
+export { PasswordResetForm } from "./PasswordResetForm";
