+++ src/shared/locales/auth.ts
@@ -1,0 +1,7 @@
+export const authTexts = {
+  enterNewPassword: "Введите новый пароль",
+  confirmNewPassword: "Повторите новый пароль",
+  changePassword: "Изменить пароль",
+  changingPassword: "Изменение пароля",
+  passwordPlaceholder: "********",
+};
